+++ federatedscope/core/data/__init__.py
@@ -1,0 +1,3 @@
+from federatedscope.core.data.base_data import StandaloneDataDict, ClientData
+
+__all__ = ['StandaloneDataDict', 'ClientData']
